# Move error handling to C code