--- a/openssl.go
+++ b/openssl.go
@@ -269,19 +269,6 @@ func base(b []byte) *byte {
 	return unsafe.SliceData(b)
 }
 
-// noescapeMkcgoErrState hides a pointer from escape analysis.
-// noescapeMkcgoErrState is the identity function but escape
-// analysis doesn't think the output depends on the input.
-// noescapeMkcgoErrState is inlined and currently compiles
-// down to zero instructions.
-// USE CAREFULLY!
-//
-//go:nosplit
-func noescapeMkcgoErrState(p **C.mkcgo_err_state) **C.mkcgo_err_state {
-	x := uintptr(unsafe.Pointer(p))
-	return (**C.mkcgo_err_state)(unsafe.Pointer(x ^ 0))
-}
-
 // newMkcgoErr creates a new error from the given mkcgo_err_state
 // and frees the state. If errst is nil, it returns nil.
 func newMkcgoErr(msg string, errst *C.mkcgo_err_state) error {