--- a/shims.h
+++ b/shims.h
@@ -1,6 +1,9 @@
 // This header file is used by the mkcgo tool to generate cgo and Go bindings for the
 // OpenSSL C API. Run "go generate ." to regenerate the bindings.
 
+#ifndef _GO_OSSL_SHIMS_H // only include this header once
+#define _GO_OSSL_SHIMS_H
+
 #include <stdlib.h> // size_t
 #include <stdint.h> // uint64_t
 
@@ -382,3 +385,5 @@ int PKCS5_PBKDF2_HMAC(const char *pass, int passlen, const unsigned char *salt, 
 
 // OBJ API
 const char *OBJ_nid2sn(int n) __attribute__((noerror));
+
+#endif // _GO_OSSL_SHIMS_H